refactor: break geocode script into modules
